requirements.txt
fastapi
uvicorn
boto3
PyYAML
pydantic
python-dotenv
